test(b): migrate commands/test_view -> e2e/test_16_view (#40)
Four tests relocated. The view command doesn't accept --config, so all
CLI invocations use cli(..., config=False). viewer_server monkeypatches
preserved to prevent real server/browser launches in tests.
